update: FINAL-43 stock history function init stockhistory page

+++ projects/server/src/controllers/stockB.js
@@ -3,6 +3,7 @@ const Sequelize = require("sequelize");
 const { Op } = db.Sequelize;
 const moment = require("moment");
 const { query } = require("express");
+const { createStockHistory } = require("../service/stock.service");
 const stockControllerB = {
   getAllStock: async (req, res) => {
     try {
@@ -109,23 +110,36 @@ const stockControllerB = {
   insertStock: async (req, res) => {
     const trans = await db.sequelize.transaction();
     try {
-      const { quantity_stock, product_id, branch_id, discount } = req.body;
-      console.log(req.body);
-      await db.Stock.create(
+      const { quantity_stock, product_id, branch_id } = req.body;
+      const input = {
+        quantity_before: 0,
+        status: "INCREMENT",
+        status_quantity: quantity_stock,
+        quantity_after: quantity_stock,
+        feature: "ADDED BY ADMIN BRANCH",
+      };
+
+      const stock = await db.Stock.create(
         {
           quantity_stock: parseInt(quantity_stock),
           product_id: parseInt(product_id),
           branch_id: parseInt(branch_id),
-          discount: parseInt(discount),
         },
         {
           transaction: trans,
         }
       );
       await trans.commit();
-      return await db.Stock.findAll().then((result) => {
-        res.send(result);
-      });
+      const isStockHistoryCreated = await createStockHistory(stock, input);
+      if (isStockHistoryCreated) {
+        return await db.Stock.findAll().then((result) => {
+          res.send(result);
+        });
+      } else {
+        return res.status(500).send({
+          message: "Error creating StockHistory.",
+        });
+      }
     } catch (err) {
       console.log(err);
       await trans.rollback();
@@ -138,22 +152,30 @@ const stockControllerB = {
   editStock: async (req, res) => {
     const trans = await db.sequelize.transaction();
     try {
-      const { quantity_stock, discount } = req.body;
+      const { quantity_stock, status, quantity_before, status_quantity } =
+        req.body;
+
+      const input = {
+        quantity_before,
+        status,
+        status_quantity,
+        quantity_after: quantity_stock,
+        feature: "UPDATED BY ADMIN BRANCH",
+      };
+
       const stok = await db.Stock.findOne({
         where: {
           id: req.params.id,
         },
       });
-      console.log(stok);
+
       const qty_stok = quantity_stock
         ? quantity_stock
         : stok.dataValues.quantity_stock;
-      const disc = discount ? discount : stok.dataValues.discount;
 
-      await db.Stock.update(
+      const stockUpdate = await db.Stock.update(
         {
           quantity_stock: qty_stok,
-          discount: disc,
         },
         {
           where: {
@@ -162,12 +184,20 @@ const stockControllerB = {
           transaction: trans,
         }
       );
+
       await trans.commit();
-      return await db.Stock.findOne({
-        where: {
-          id: req.params.id,
-        },
-      }).then((result) => res.send(result));
+      const isStockHistoryCreated = await createStockHistory(stok, input);
+      if (isStockHistoryCreated) {
+        return await db.Stock.findOne({
+          where: {
+            id: req.params.id,
+          },
+        }).then((result) => res.send(result));
+      } else {
+        return res.status(500).send({
+          message: "Error creating StockHistory.",
+        });
+      }
     } catch (err) {
       console.log(err.message);
       await trans.rollback();
@@ -180,15 +210,37 @@ const stockControllerB = {
   deleteStock: async (req, res) => {
     const trans = await db.sequelize.transaction();
     try {
-      await db.Stock.destroy({
+      const { quantity_before } = req.query;
+      const input = {
+        quantity_before,
+        status: "DECREMENT",
+        status_quantity: quantity_before,
+        quantity_after: 0,
+        feature: "DELETED BY ADMIN BRANCH",
+      };
+
+      const stok = await db.Stock.findOne({
         where: {
-          //  id: req.params.id
-          //   [Op.eq]: req.params.id
           id: req.params.id,
         },
       });
-      await trans.commit();
-      return await db.Stock.findAll().then((result) => res.send(result));
+
+      const isStockHistoryCreated = await createStockHistory(stok, input);
+      if (isStockHistoryCreated) {
+        await db.Stock.destroy({
+          where: {
+            //  id: req.params.id
+            //   [Op.eq]: req.params.id
+            id: req.params.id,
+          },
+        });
+        await trans.commit();
+        return await db.Stock.findAll().then((result) => res.send(result));
+      } else {
+        return res.status(500).send({
+          message: "Error creating StockHistory.",
+        });
+      }
     } catch (err) {
       console.log(err.message);
       await trans.rollback();
@@ -201,7 +253,6 @@ const stockControllerB = {
   getAllStockByCategory: async (req, res) => {
     try {
       const { category_name } = req.query;
-      console.log(req.query);
       const get = await db.Stock.findAll({
         include: [
           {
@@ -254,6 +305,50 @@ const stockControllerB = {
       return res.send(product);
     } catch (err) {
       console.log(err.message);
+      res.status(500).send({
+        message: err.message,
+      });
+    }
+  },
+  getStockHistory: async (req, res) => {
+    try {
+      const get = await db.StockHistory.findAll({
+        include: [
+          {
+            model: db.Stock,
+            as: "Stock",
+            attributes: ["id", "quantity_stock"],
+            include: [
+              {
+                model: db.Product,
+                as: "Product",
+                attributes: [
+                  "product_name",
+                  "price",
+                  "photo_product_url",
+                  "category_id",
+                  "desc",
+                  "weight",
+                ],
+                include: [
+                  {
+                    model: db.Category,
+                    as: "Category",
+                    attributes: ["category_name"],
+                  },
+                ],
+              },
+              {
+                model: db.Branch,
+                as: "Branch",
+                attributes: ["id", "branch_name"],
+              },
+            ],
+          },
+        ],
+      });
+      res.send(get);
+    } catch (err) {
       res.status(500).send({
         message: err.message,
       });
